Scripts/Karazhan: - nightbane non attackable while flying - aran fixed in middle - arcane immunity for arcane mobs - no more dance vibe
Core/OPvP:
- tryfix spawns for halaa - not yet tested

Core/Spells:
- fix master totems

+++ src/game/OutdoorPvPNA.cpp
@@ -97,7 +97,7 @@ void OPvPCapturePointNA::SpawnNPCsForTeam(uint32 team)
         creatures=HordeControlNPCs;
     else
         return;
-    for (int i = 0; i < NA_CONTROL_NPC_NUM; ++i)
+    for (uint8 i = 0; i < NA_CONTROL_NPC_NUM; ++i)
         AddCreature(i,creatures[i].entry,creatures[i].teamval,creatures[i].map,creatures[i].x,creatures[i].y,creatures[i].z,creatures[i].o,1000000);
 }
